[Add] 2_api_resources - DeleteSongsTest - just_authorized_artists_can_delete_a_song

+++ 2_api_resources/app/Http/Controllers/SongController.php
@@ -55,6 +55,11 @@ class SongController extends Controller
 
     public function destroy(Song $song)
     {
+        if (! auth()->user()->can('delete', $song)) {
+            return response()->json([
+                'error' => 'You are not allowed to perform this action'
+            ], 403);
+        }
         $song->delete();
 
         return new SongResource($song);
